fake voting UI linking to backend

+++ lib/domain/entities/vote.dart
@@ -1,6 +1,6 @@
 class Vote{
-   double fakeVotes;
-   double originalVotes;
+   int fakeVotes;
+   int originalVotes;
 
    Vote({required this.fakeVotes, required this.originalVotes});
